move v1 db connection out of v1 job service`
