+++ dpc-go/dpc-attribution/src/main.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"net/http"
 
+	"github.com/CMSgov/dpc/attribution/service"
 	"go.uber.org/zap"
 
 	"github.com/CMSgov/dpc/attribution/conf"
@@ -33,7 +34,7 @@ func main() {
 	or := repository.NewOrganizationRepo(db)
 	os := v2.NewOrganizationService(or)
 
-	js := v1.NewJobService(ctx)
+	js := createV1Services(logger.WithContext(ctx))
 	gr := repository.NewGroupRepo(db)
 	gs := v2.NewGroupService(gr, js)
 
@@ -47,4 +48,20 @@ func main() {
 	if err := http.ListenAndServe(fmt.Sprintf(":%s", port), attributionRouter); err != nil {
 		logger.WithContext(ctx).Fatal("Failed to start server", zap.Error(err))
 	}
+}
+
+func createV1Services(log *zap.Logger) service.JobService {
+	attrDbV1 := repository.GetAttributionV1DbConnection()
+	queueDbV1 := repository.GetQueueDbConnection()
+	defer func() {
+		if err := attrDbV1.Close(); err != nil {
+			log.Fatal("Failed to close attribution v1 db connection", zap.Error(err))
+		}
+		if err := queueDbV1.Close(); err != nil {
+			log.Fatal("Failed to close queue v1 db connection", zap.Error(err))
+		}
+	}()
+	pr := repository.NewPatientRepo(attrDbV1)
+	jr := repository.NewJobRepo(queueDbV1)
+	return v1.NewJobService(pr, jr)
 }
